Move SparkplugDevice to a new file.
